[feat]#1 - 게시글 & 댓글 CRUD # 📝 본문 (Commit Body) - 게시글과 댓글 연결 - ApiResponse을 통해 API 응답 일관화 시킴. - ResponseEntity 적용
# 기타 참고 사항
- 관련 이슈: #1

+++ src/main/java/com/tev/tev/board/controller/BoardController.java
@@ -1,10 +1,16 @@
 package com.tev.tev.board.controller;
 
-import com.tev.tev.board.dto.BoardCreate;
-import com.tev.tev.board.dto.BoardEdit;
+import com.tev.tev.board.dto.request.BoardRequest;
+import com.tev.tev.board.dto.response.BoardListResponse;
+import com.tev.tev.board.dto.response.BoardResponse;
+import com.tev.tev.board.dto.request.BoardUpdate;
 import com.tev.tev.board.entity.Board;
+import com.tev.tev.board.repository.BoardRepository;
 import com.tev.tev.board.service.BoardService;
+import com.tev.tev.common.ApiResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,44 +21,72 @@ import java.util.List;
 public class BoardController {
 
     private final BoardService boardService;
-    // 파라미터로 변경
+    private final BoardRepository boardRepository;
 
     // 게시글 생성
     @PostMapping("/create")
-    public String createdBoard(BoardCreate boardCreate){
-        boardService.create(boardCreate);
-        return "게시글 생성 완료";
+    public ResponseEntity<ApiResponse<String>> boardCreate(BoardRequest boardRequest){
+        Integer boardId = boardService.saveBoard(boardRequest);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(ApiResponse
+                        .success("게시글 생성 성공 id:" + boardId));
     }
 
     // 게시글 모두 조회
     @GetMapping("/list")
-    public List<Board> listBoard(){
-        return boardService.list();
+    public ResponseEntity<ApiResponse<List<BoardListResponse>>> boardList(){
+        List<BoardListResponse> boardList = boardService.getBoardList();
+        if(boardList.isEmpty()){
+            return ResponseEntity
+                    .ok(ApiResponse.success("게시글이 존재하지 않습니다.", null));
+        }
+        return ResponseEntity
+                .ok(ApiResponse.success(boardList));
     }
 
     // 게시글 조회 - id
-    @GetMapping("/view/{boardid}")
-    public Board viewBoard(@PathVariable("boardid") Integer boardId){
-        return boardService.view(boardId);
+    @GetMapping("/{boardid}")
+    public ResponseEntity<ApiResponse<BoardResponse>> viewBoard(@PathVariable("boardid") Integer boardId){
+        if(!boardRepository.existsById(boardId)){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(ApiResponse.fail("존재하지 않는 게시글입니다. id: " + boardId));
+        }
+
+        BoardResponse boardResponse = boardService.getBoardById(boardId);
+        return ResponseEntity
+                .ok(ApiResponse.success("게시글 조회 성공", boardResponse));
     }
 
-    // 유저가 작성한 게시글 조회
-    @GetMapping("/list/{userid}")
-    public List<Board> userListBoard(@PathVariable("userid") Integer userId){
-        return boardService.userList(userId);
-    }
+    // TODO: 유저가 작성한 게시글 조회
 
     // 게시글 수정
-    @PatchMapping("/edit")
-    public String editBoard(@RequestParam("boardid") Integer boardId, BoardEdit boardEdit){
-        boardService.edit(boardId, boardEdit);
-        return "게시글 수정 완료 id: " + boardId;
+    @PatchMapping("/{boardid}/edit")
+    public ResponseEntity<ApiResponse<BoardResponse>> editBoard(@PathVariable("boardid") Integer boardId, BoardUpdate boardUpdate){
+        if(!boardRepository.existsById(boardId)){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(ApiResponse.fail("존재하지 않는 게시글입니다. id: " + boardId));
+        }
+
+        BoardResponse boardResponse = boardService.updateBoard(boardId, boardUpdate);
+        if(boardResponse == null){
+            BoardResponse findBoard = boardService.getBoardById(boardId);
+            return ResponseEntity
+                    .ok(ApiResponse.success("수정된 사항이 없습니다.", findBoard));
+        }
+
+        return ResponseEntity
+                .ok(ApiResponse.success("게시글 수정 성공", boardResponse));
     }
 
-    @DeleteMapping("/delete")
-    public String deleteBoard(@RequestParam("boarid") Integer boardId){
+    @DeleteMapping("/{boardid}/delete")
+    public ResponseEntity<ApiResponse<String>> deleteBoard(@PathVariable("boarid") Integer boardId){
+        if(!boardRepository.existsById(boardId)){
+            return ResponseEntity.badRequest()
+                    .body(ApiResponse.fail("존재하지 않는 게시글입니다. id: " + boardId));
+        }
+
         boardService.delete(boardId);
-        return "게시글 삭제 완료 id: " + boardId;
+        return ResponseEntity.ok(ApiResponse.success("게시글 삭제 성공 id: " + boardId));
     }
